Switch to 64 bit ints by default, convert 64 ints to number in TS

+++ lang/rs/project/src/project/mod.rs
@@ -469,7 +469,7 @@ use crate::named::Name;
 pub struct Admin {
     pub name: Name,
     pub email: String,
-    pub age: Option<i32>,
+    pub age: Option<i64>,
     pub role: AdminRole,
 }
 
@@ -514,7 +514,7 @@ use serde::{Deserialize, Serialize};
 pub struct User {
     pub name: Name,
     pub email: String,
-    pub age: Option<i32>,
+    pub age: Option<i64>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
